Add support for sending data by UDP in addition to logging
This allows you to log data remotely by sending UDP to another
server, or have some other process act on the data without
watching a moving target (the file being logged to).

The impetus for this was echoing log entries to IRC, hence the
option to apply IRC colour codes.

Resolves GH #5